image_text_to_word/requirements.txt
python-docx>=0.8.11
Pillow>=9.0.0